Tidied up main()
The whiteArray generator and average value calculator portions of the main code were moved out to a separate function, reducing the number of lines for main to 78 lines. This can still be further reduced. I've kept some lines/variables in main as they may still be crucial to finishing challenge.